+++ src/main/java/com/minhcraft/beyondbetatweaks/mixin/feature/enchantment/glint/ItemMixin.java
@@ -1,6 +1,5 @@
-package com.minhcraft.beyondbetatweaks.mixin.item;
+package com.minhcraft.beyondbetatweaks.mixin.feature.enchantment.glint;
 
-import net.minecraft.world.item.Equipable;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import org.spongepowered.asm.mixin.Mixin;
@@ -9,14 +8,12 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(Item.class)
-public abstract class ArmorMixin {
+public abstract class ItemMixin {
 
-    // Disable enchantment glint on enchanted armor
+    // Disable enchantment glint on enchanted armor and tools
     // Code copied from https://github.com/jmb05/LessGlintyThings
     @Inject(method = "isFoil", at = @At("HEAD"), cancellable = true)
-    private void beyond_beta_tweaks$disableEnchantedArmorGlint(ItemStack stack, CallbackInfoReturnable<Boolean> cir) {
-        if (stack.getItem() instanceof Equipable) {
-            cir.setReturnValue(false);
-        }
+    private void beyond_beta_tweaks$disableEnchantmentGlint(ItemStack stack, CallbackInfoReturnable<Boolean> cir) {
+        cir.setReturnValue(false);
     }
 }